src/requirements.txt
torch
numpy
plotly
matplotlib
fake-bpy-module-4.3
